Replaced the contents of the fed config by fedtest
More than 6 years ago the work was started to replace FED by something
new. This was done by placing a second configuration next to FED, called
FEDTest and instruction the shifters (already in 2009) to look at
FEDTest instead of FED. In the meanwhile FED was completely broken and
FEDTest was partly broken. FEDTest was now fixed in #12240.
We placed the contents of FEDTest into FED so that this can be the main
default plot to look at again. We should stop running FEDTest in the
near future and adapt the instructions at that moment.